ark_list_web_service/requirements.txt
Flask>=2.3,<3.0
requests>=2.32,<3.0
llama-cpp-python==0.2.87
yahooquery>=2.3,<3.0
